requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.2
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
install==1.3.5
mongoengine==0.23.1
pycparser==2.21
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4
typing_extensions==4.0.1
whitenoise==6.0.0
wincertstore==0.2